refactor: Ktlint 적용
