+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
         android:allowBackup="true"
@@ -24,8 +23,7 @@
         </activity>
         <activity
             android:name=".ui.addcard.AddCardActivity"
-            android:exported="true"
-            />
+            android:exported="true" />
     </application>
 
 </manifest>
