+++ app/src/main/java/woowacourse/payments/ui/addcard/AddCardScreen.kt
@@ -2,28 +2,24 @@ package woowacourse.payments.ui.addcard
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.ui.addcard.component.AddCardTopbar
-import woowacourse.payments.ui.component.Card
 import woowacourse.payments.ui.addcard.component.CardNumberTextField
 import woowacourse.payments.ui.addcard.component.ExpireDateTextField
 import woowacourse.payments.ui.addcard.component.OwnerNameTextField
 import woowacourse.payments.ui.addcard.component.PasswordTextField
+import woowacourse.payments.ui.component.Card
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
+import woowacourse.payments.ui.uimodel.CardInfoUiState
 
 @Composable
 fun AddCardScreen(cardInfo: CardInfoUiState, modifier: Modifier = Modifier) {
@@ -70,8 +66,9 @@ private fun AddCardScreenPreview() {
                 AddCardTopbar()
             }
         ) {
-            AddCardScreen(CardInfoUiState(), modifier = Modifier
-                .padding(it)
+            AddCardScreen(
+                CardInfoUiState(), modifier = Modifier
+                    .padding(it)
             )
         }
     }
