+++ app/src/main/java/woowacourse/payments/ui/allcards/util/CardFormatter.kt
@@ -1,7 +1,5 @@
 package woowacourse.payments.ui.allcards.util
 
-import androidx.compose.ui.text.AnnotatedString
-
 object CardFormatter {
     fun formatCardNumber(text: String): String = text.chunked(4)
         .mapIndexed { index, s ->
